RavenDB-25281 WP8 cleanup: consolidate CoraxQueryBuilder, wire heuristics, delete dead code
- Consolidated CoraxQueryBuilder (3 partial files) into QueryPlanBuilder
  * VectorSearch logic → QueryPlanBuilder.HandleVector + ResolveVectorItems
  * Parameters → standalone QueryBuilderParameters.cs
  * Main → CoraxIndexReadOperation for MoreLikeThis path

- IL emitter now calls QueryPrimitives.ShouldSwitchToEntryScan(long, long)
  instead of inlining magic numbers (32000, 64). Single source of truth for
  entry-scan cost thresholds via named constants.

- Deleted dead code never wired up:
  * Primitives.cs: ScanAndFilter, ScanWithSkipAndLimitOnly
  * IndexSearcher.cs: BatchGetEntryData, CreateEntryTermsReader
  * PlanOpKind enum: 9 dead values (FillFromRange, ScanAndFilter*, etc.)
  * Unused constants: SkipBufferSize, FieldRootPagesStackAllocThreshold

- Kept (.Erasure.cs files, IQueryMatch, CoraxVectorItem) — still in use.

All 654 Corax tests pass. Build clean.
