+++ src/Corax/Querying/Planning/QueryPlan.cs
@@ -10,16 +10,7 @@ public enum PlanOpKind : byte
     AndNotWithPostings,
     LazyOrWithPostings,
     RepairAfterLazy,
-    FillFromRange,
     CheckAndMaybeEntryScan,
-    ScanAndFilter,
-    ScanAndFilterInPlace,
-    SortWithFilter,
-    OrderedRangeScan,
-    VectorRank,
-    SpatialFilter,
-    SortByScore,
-    SortByDistance,
     IterateInto,
     DirectIterate,
     /// <summary>Clear a specific bitmap slot. BitmapLocal = slot index.</summary>
